perf(runner): merge vm-init and vsock-agent into single rust binary
Create new vm-init crate that combines:
- Filesystem initialization (mounts, overlayfs, pivot_root)
- PID 1 responsibilities (signal handling, zombie reaping)
- vsock-agent IPC functionality

Changes:
- Add crates/vm-init with init.rs, pid1.rs, main.rs
- Refactor vsock-agent into library (lib.rs) + binary (main.rs)
- Update vm-init.sh to exec the new vm-init binary
- Update build-rootfs.sh to copy vm-init instead of vsock-agent
- Update deploy-runner.yml ansible playbook
- Update CI workflows to build vm-init for ARM64
- Remove tini from Dockerfile (no longer needed)
- Update comments in vm.ts

This eliminates the need for tini and the shell script overhead,
reducing VM boot time by ~10-15ms.

Closes #1807
