+++ .github/workflows/turbo.yml
@@ -813,13 +813,13 @@ jobs:
       - name: Build Core Package
         run: cd turbo && pnpm turbo run build --filter=@vm0/core
 
-      # Cross-compile Rust vsock-agent for ARM64 (metal runners are ARM64)
-      - name: Build vsock-agent for ARM64
+      # Cross-compile Rust vm-init for ARM64 (metal runners are ARM64)
+      - name: Build vm-init for ARM64
         run: |
           cd crates
           CARGO_TARGET_AARCH64_UNKNOWN_LINUX_MUSL_LINKER=aarch64-linux-gnu-gcc \
-            cargo build --release --target aarch64-unknown-linux-musl -p vsock-agent
-          ls -la target/aarch64-unknown-linux-musl/release/vsock-agent
+            cargo build --release --target aarch64-unknown-linux-musl -p vm-init
+          ls -la target/aarch64-unknown-linux-musl/release/vm-init
 
       - name: Build runner bundle
         run: |
