+++ crates/vm-init/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "vm-init"
+version = "0.1.0"
+edition = "2024"
+description = "VM init process for Firecracker - filesystem setup, PID 1, and vsock-agent"
+
+[dependencies]
+libc = "0.2"
+nix = { version = "0.29", default-features = false, features = ["mount", "fs"] }
+vsock-agent = { path = "../vsock-agent" }
+
+[lints]
+workspace = true
